Large commit incoming.
Add mysql service type to get uptime of a mysql server.
Integrate Humanizer library (project reference).
Add channel restrictions to services (services can only be shown to channels that are listed in the config file)
Add probing path for humanizer lib so we don't see a mess in our cwd()